refactor: useModal custom hook 추가, 관련 로직 수정
